Created p13.py
Super easy. Literally add up all the numbers in the list and look at only the last ten digits. This is only possible because of python. In other languages such as Java or C++, you would have to resort to using a big number library function or, gasp, actually get clever.