Created p19.py
The logic at times is tough to follow, but is simple at it's heart. All it is, is noticing that there is cyclical patterns on how many Sundays fall on the first of each month and the logic is merely expanding and deriving all the relationships. At the end it is shown that there are 171 Sundays falling on the first of the month in that period.